lb 5

+++ lb5/src/Criterion.java
@@ -1,9 +1,10 @@
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
 public class Criterion {
     String criterionName;
-    double criterionMark;
+    BigDecimal criterionMark;
     int criterionNum;
 
     List<String> criterionValues = new ArrayList<>();
